+++ Backend/Database/ds.js
@@ -1,15 +1,6 @@
 // Importing required modules
 const mongoose = require("mongoose");
-const express = require("express");
 require("dotenv").config(); // Loading environment variables from .env file
-const bcrypt = require("bcrypt");
-const jwt = require('jsonwebtoken')
-
-// Creating an instance of Express Router
-const MDrouter = express.Router();
-
-// Importing Mongoose models
-const { UserModel } = require("../modules/MDSchema");
 
 // Function to start the database connection
 const startDatabase = async () => {
@@ -41,44 +32,9 @@ const isConnected = () => {
   return mongoose.connection.readyState === 1;
 };
 
-const validateEmail = async (email) => {
-  try {
-    // Check if there is a user with the provided email
-    const user = await UserModel.findOne({ email: email });
-    // If user exists, return true; otherwise, return false
-
-    return !!user;
-  } catch (error) {
-    // Handle error
-    console.error("Error validating email:", error.message);
-    // Return false in case of error
-    return false;
-  }
-};
-
-const validatePassword = async (pass, email) => {
-  try {
-    // Check if there is a user with the provided email
-    const user = await UserModel.findOne({ email: email });
-    console.log(user);
-    // Use bcrypt's compare function to compare the plain password with the hashed password
-    const match = await bcrypt.compare(pass, user.password);
-
-    // Return true if the password matches the hashed password, false otherwise
-    return match;
-  } catch (error) {
-    console.error(error);
-    // Handle any errors gracefully
-    return false;
-  }
-};
-
 // Exporting functions and router for use in other modules
 module.exports = {
   startDatabase,
   stopDatabase,
-  isConnected,
-  validateEmail,
-  validatePassword,
-  MDrouter,
+  isConnected
 };
